AG-3390 - Extract `urlWithBaseUrl` and `getPathFromUrlPathname` from pages
So it can be used on the client side
