+++ packages/ag-charts-website/src/components/site-header/SiteHeader.astro
@@ -4,10 +4,11 @@ import { HeaderNav } from './HeaderNav';
 import { getEntry } from 'astro:content';
 import { FRAMEWORKS } from '@constants';
 import gridStyles from './gridSiteHeader.module.scss';
-import { urlWithBaseUrl } from '@utils/pages';
 import { SiteLogo } from './SiteLogo';
 import Search from '../search/Search';
 import { getFrameworkFromPath } from '../../features/docs/utils/urlPaths';
+import { urlWithBaseUrl } from '@utils/urlWithBaseUrl';
+
 const { pathname } = Astro.url;
 const { data: menuData } = await getEntry('menu', 'data');
 const topNavItems = menuData.header.items;
