+++ packages/ag-charts-website/src/layouts/Layout.astro
@@ -5,7 +5,7 @@ import SiteHeader from '@components/site-header/SiteHeader.astro';
 import '@design-system/design-system.scss';
 import { Footer } from '@components/footer/Footer';
 import type { ApiMenuItem } from '@ag-grid-types';
-import { urlWithBaseUrl } from '@utils/pages';
+import { urlWithBaseUrl } from '@utils/urlWithBaseUrl';
 import { ApiTopBar } from '@components/top-bar/ApiTopBar';
 import { getIsProduction } from '@utils/env';
