+++ packages/ag-charts-website/src/utils/urlWithBaseUrl.ts
@@ -1,0 +1,8 @@
+import { SITE_BASE_URL } from '../constants';
+
+export const urlWithBaseUrl = (url: string = '') => {
+    const regex = /^\/(.*)/gm;
+    const substitution = `${SITE_BASE_URL}$1`;
+
+    return url.match(regex) ? url.replace(regex, substitution) : url;
+};
